PLAD-67 - As a user, I want to be able to log in, so that I have access to the system (#55)
* [Frontend] - Drawer refactoring

* [Frontend] - Login page implementation

* [Frontend] - Location initial value fix

* [Frontend] - Removed useless imports

* [Frontend] - JWT handling and http wrapper

* [BACKEND] Added login mechanism, added seeds for registration, added migrations for user_id columns

* [BACKEND] Fixed student property on user profile

* [Backend] - Fixed minor bugs

* [Frontend] - Improved login services

* [Frontend] - Improved login screen

* [pubspec.yaml] - Updated sharedpreferences version

* [Frontend] - Working login flow up to create placement and student

* [Frontend] - App flow improvements, fixed minor location bugs

* [Frontend] - Fixed routing and improved routing generator

* [Frontend] - Added userId to placement and student models and creation

* [BACKEND] Added general authorization middleware on all urls besides login and registration

* [BACKEND] Added route authorization for half of the route files, where needed

* [BACKEND] Fixed get conversation only by one of the 2 logged in users

* [BACKEND] Secured creation of a message

* [BACKEND] Updated the delete messages route

* [BACKEND] Secured save match route that was left out

* [BACKEND] Added studentId and employerId in the login token

* [BACKEND] Added changes to routes and services after adding the student id and the employer id in the JWT

* [BACKEND] Secured placement routes

* [BACKEND] Secured recommendations endpoints

* [BACKEND] Secured student routes

* [BACKEND] Fixed some errors in the documentation, added authorization

* [BACKEND] Secured messages route further, and fixed messages documentation

* [BACKEND] Updated the documentation with the latest authorization issues

* [BACKEND] Fixed registration and login tests

* [BACKEND] Added a new seed file for matches

* [BACKEND] Fixed tests for matching to support authentication

* [BACKEND] Fixed messages tests to support auth

* [BACKEND] Fixed tests for placements with auth

* [BACKEND] Fixed tests for recommendations

* [BACKEND] Fixed skills tests with auth

* [BACKEND] Fixed student tests with authentication; fixed saving of student profile from tests from hardcoded IDs to dynamic alocation

* [BACKEND] Updated seeds after merge

* [BACKEND] Fixed seeds; fixed new routes and tests to adhere to authentication

* [BACKEND] - Fast fixes

* [FRONTEND] - Fixed location bug on profile

* [FRONTEND] - Fixes after merges

* [FRONTEND] - Login bug fixing

* [FRONTEND] - Fast fixes

* [Frontend] - DomainOfActivity dropdown fix

* [Frontend] - Employer app flow fixes

* [BACKEND] Routes and seeds review


* [frontend] - Fixed edge cases in company recommendations

* [Frontend] - Default api_services fallback case

* [Frontend] - Dynamic chat names

* [Frontend] - Deleted TODOs

* [Frontend] - Chat screen color fixes

* [FRONTEND] - Bugfixing after merges

* [Frontend] - minor chat ui fixes

* [Frontend] - Dynamic values in match alert

* [Frontend] - Calendar UI colors fix

* [Frontend] - Login fields UI fix

* [FRONTEND] - Scrollable skill box

* [FRONTEND] - Fixed create student

* [FRONTEND] - Fixed create placement

* [FRONTEND] - Fixed create employer

* [Frontend] - Company student card dynamic fields

* [Frontend] - Student placement card dynamic fields

* [Frontend] - Employer screen UI fixes

* [Frontend] - Removed TODO

* [BACKEND] Fixed failing test

* [Frontend] - Multiselect improvements

* [Frontend] - Padding in recommendation lists

* [BACKEND] Refresh JWT on student or employer profile creation

* [backend] - seed refactoring

* [bakend] - fix to the deletion of email and password for the student

* [Frontend] - Skills recommendation fix

* [Backend] - crashing s.email field not needed

* [Frontend] - Minor fixes

* [backend] - small fix to the algorithm

* [Frontend] - Minor JWT token handling bugfix

* [Frontend] - Bug fixing

* [BACKEND] Fixed TODOs

* [BACKEND] Fixed unique email address on registration

* [BACKEND] Fixes after code review

* [BACKEND] Revert the employer authorization verification and fixed get placement by id

* [FRONTEND] - Fixed UI drawer

* [FRONTEND] - Added registration error handling

* [FRONTEND] -  Fixed employer creation UI

* [FRONTEND] - Fixed remove placement after filtering

* [FRONTEND] - Changed type of dropdown builder in employer student list

* [Frontend] - Reverting previous commit

+++ backend/services/employerService.js
@@ -2,10 +2,11 @@ const employerDAO = require('../DAO/employerDAO');
 
 const SuperError = require('../errors').SuperError;
 const ERR_BAD_REQUEST = require('../errors').ERR_BAD_REQUEST;
+const ERR_FORBIDDEN = require('../errors').ERR_FORBIDDEN;
 
 const locationService = require('../services/locationService');
 const domainOfActivityService = require('../services/domainOfActivityService');
-
+const jwt = require('jsonwebtoken');
 
 
 module.exports = employerService = {
@@ -13,15 +14,25 @@ module.exports = employerService = {
         return employerDAO.getEmployer(id);
     },
 
-    getLastEmployer: async () => {
-        return await employerDAO.getLastEmployer();
+    getEmployerByUserId: (userId) => {
+        return employerDAO.getEmployerByUserId(userId);
     },
-
-    createNewEmployer: async (details) => {
+    getEmployerByPlacementId: (placementId) => {
+        return employerDAO.getEmployerByPlacementId(placementId);
+    },
+    createNewEmployer: async (details, auth) => {
         if(details && typeof(details.location.country) == 'string' && typeof(details.location.city) == 'string' && typeof(details.name) == 'string' && typeof(details.description) == 'string' && !isNaN(details.domainOfActivityId)){
+            if(auth.userType !== 'EMPLOYER') {
+                throw new SuperError(ERR_FORBIDDEN, 'You cannot create an employer profile from a non-employer account.')
+            }
             let check = await domainOfActivityService.existsDomainOfActivity(details.domainOfActivityId);
             if(check){
-                let newEmployer = await employerDAO.addNewEmployer(details);
+                let newEmployer = await employerDAO.addNewEmployer(details, auth.id);
+                if(newEmployer.id) {
+                    newEmployer.token = jwt.sign({ id: auth.id, employerId: newEmployer.id, userType: 'EMPLOYER'}, process.env.ACCESS_TOKEN_SECRET, {
+                        expiresIn: process.env.ACCESS_TOKEN_LIFE // 30 DAYS
+                    });
+                }
                 if(details.location){
                     newEmployer.location = await employerService.saveEmployerLocation(newEmployer.id, details.location);
                 }
@@ -43,7 +54,10 @@ module.exports = employerService = {
         return location;
     },
 
-    deleteEmployerById: (id) => {
-        return employerDAO.deleteEmployerById(id);
+    deleteEmployerById: async (id, auth) => {
+        if(auth.employerId !== id) {
+            throw new SuperError(ERR_FORBIDDEN, 'You cannot delete this employer profile');
+        }
+        return await employerDAO.deleteEmployerById(id);
     }
 };
