+++ backend/services/matchService.js
@@ -1,8 +1,24 @@
 const matchDAO = require('../DAO/matchDAO');
+const { checkIfOtherSkillExists } = require('../DAO/skillsDAO');
+const { SuperError, ERR_FORBIDDEN, ERR_NOT_FOUND } = require('../errors');
+const employerService = require('./employerService');
+const placementService = require('./placementService');
 const studentService = require('./studentService');
 
 module.exports = {
-    saveChoice: async (choice) => { 
+    saveChoice: async (choice, auth) => { 
+        let employer = await employerService.getEmployerByPlacementId(choice.placementID);
+        if (auth.studentId !== choice.studentID && auth.employerId !== employer.id) {
+            throw new SuperError(ERR_FORBIDDEN, 'You are not authorized to save this interaction');
+            return;
+        }
+        if(auth.studentId && choice.placementAccept !== undefined) {
+            throw new SuperError(ERR_FORBIDDEN, 'You are not allowed to save this interaction'); return;
+        }
+        if(auth.employerId && choice.studentAccept !== undefined) {
+            throw new SuperError(ERR_FORBIDDEN, 'You are not allowed to save this interaction'); return;
+        }
+
         let previousInteraction = await matchDAO.getPreviousInteraction(choice.studentID, choice.placementID);
         let result = {};
         if(!previousInteraction) {
@@ -13,15 +29,35 @@ module.exports = {
         return result;
     },
 
-    getMatchesByStudentId: (studentId) => {
+    getMatchesByStudentId: (studentId, auth) => {
+        if(parseInt(studentId) !== auth.studentId) {
+            throw new SuperError(ERR_FORBIDDEN, 'You are not authorized to retrieve this information');
+            return;
+        }
         return matchDAO.getMatchesByStudentId(studentId);
     },
     
-    deleteMatch: async (studentId, placementId) => {
-        return await matchDAO.deleteMatch(studentId, placementId);
+    deleteMatch: async (studentId, placementId, auth) => {
+        // let student = await studentService.getStudent(studentId);
+        let employer = await employerService.getEmployerByPlacementId(placementId);
+        if (auth.studentId !== studentId && auth.employerId !== employer.id) {
+            throw new SuperError(ERR_FORBIDDEN, 'You are not authorized to delete this match');
+            return;
+        }
+        return await matchDAO.deleteMatch(studentId, placementId);  
     },
 
-    getMatchesByPlacementId: async (placementId) => {        
+    getMatchesByPlacementId: async (placementId, auth) => {        
+        let employer = await employerService.getEmployerByPlacementId(placementId);
+        if(employer === null) {
+            throw new  SuperError(ERR_NOT_FOUND, 'The placement or the employer were not found');
+            return;
+        }
+        if (auth.id !== employer.userId && auth.employerId !== employer.id) {
+            throw new SuperError(ERR_FORBIDDEN, 'You are not authorized to get this information');
+            return;
+        }
+
         let matches = await matchDAO.getMatchesByPlacementId(placementId);
         
         let niceMatches = [];
@@ -46,5 +82,9 @@ module.exports = {
             }
         });
         return niceMatches;
+    },
+    checkMatch: async function(studentId, employerId) {
+        let interaction = await matchDAO.getStudentMatchesWithEmployer(employerId,studentId);
+        return (interaction.length > 0);
     } 
 };
