+++ backend/services/placementService.js
@@ -7,22 +7,19 @@ const locationService = require('./locationService');
 const matchService = require('./matchService');
 const SuperError = require('../errors').SuperError;
 const ERR_INTERNAL_SERVER_ERROR = require('../errors').ERR_INTERNAL_SERVER_ERROR;
+const ERR_FORBIDDEN = require('../errors').ERR_FORBIDDEN;
 const ERR_NOT_FOUND = require('../errors').ERR_NOT_FOUND;
 
 
 module.exports = {
 
-    
-    getAllPlacementsIds: async () => {
+    savePlacement: async (placementDetails, auth) => {
 
-        return await placementDAO.getAllPlacementsIds();
-
-    },
-
-    savePlacementPage: async (placementDetails) => {
+        if(auth.employerId !== placementDetails.employerId) {
+            throw new SuperError(ERR_FORBIDDEN, 'You are not allowed to save a placement of a different employer')
+        }
 
         let newPlacement = {};
-
         try {
 
             newPlacement = await placementDAO.createNewPlacement(placementDetails);
@@ -59,11 +56,7 @@ module.exports = {
                     newSkills = [...newSkills, ...otherSkills];  
                 }
                 newPlacement.skills = await placementDAO.setPlacementSkills(newPlacement.id, newSkills);
-
             }
-
-            
-
         } catch(error) {
             if(!error.code) {
                 throw new SuperError(ERR_INTERNAL_SERVER_ERROR, error.message || "Server Error");
@@ -74,36 +67,56 @@ module.exports = {
         return newPlacement;
     },
 
-  
-
     getPlacementsForRecommendations: async (studentId) => {
         return await placementDAO.getPlacementsForRecommendations(studentId);
     },
 
     getPlacementById: async (placementId) => {
-
-        let placement = await placementDAO.getPlacementById(placementId);
-        placement.institutions = await placementDAO.getPlacementInstitutions(placementId);
-        placement.majors = await placementDAO.getPlacementMajors(placementId);
-        placement.skills = await skillsService.getPlacementSkills(placementId);
-        placement.employer = await employerService.getEmployer(placement.employer_id);
-        placement.location = await  placementDAO.getPlacementLocation(placementId);
+        let placement = {};
+        try {
+            placement = await placementDAO.getPlacementById(placementId);
+            if(!placement) {
+                throw new SuperError(ERR_NOT_FOUND, 'The placement you are trying to see does not exist.')
+            }
+            placement.institutions = await placementDAO.getPlacementInstitutions(placementId);
+            placement.majors = await placementDAO.getPlacementMajors(placementId);
+            placement.skills = await skillsService.getPlacementSkills(placementId);
+            placement.employer = await employerService.getEmployer(placement.employer_id);
+            placement.location = await  placementDAO.getPlacementLocation(placementId);
+        } catch(error) {
+            if(error.code) {
+                throw error;
+            } else {
+                throw new SuperError(ERR_INTERNAL_SERVER_ERROR, 'There was a problem retrieving this placement. Please try again.')
+            }
+        }
         return placement;
     },
 
-    getPlacementsByEmployerId: (employerId) => {
-        return placementDAO.getPlacementsByEmployerId(employerId);
+    getEmployerByPlacementId: (placementId) => {
+        return placementDAO.getEmployerByPlacementId(placementId);
     },
 
-    deletePlacementById: (id) => {
+    getPlacementsByEmployerId: async (employerId, auth) => {
+        if(auth.employerId !== employerId) {
+            throw new SuperError(ERR_FORBIDDEN, 'You are not allowed to view the list of placements from another employer.')
+        }
+        return await placementDAO.getPlacementsByEmployerId(employerId);
+    },
+
+    deletePlacementById: async (id, auth) => {
+        let placement = await placementDAO.getPlacementById(id);
+        if(auth.employerId !== placement.employer_id) {
+            throw new SuperError(ERR_FORBIDDEN, 'You are not allowed to delete this placement')
+        }
         return placementDAO.deletePlacementById(id);
     },
 
-    getLastPlacement: () => {
-        return placementDAO.getLastPlacement();
-    },
-
-    closePlacementById: async (id) => {
+    closePlacementById: async (id, auth) => {
+        let placement = await placementDAO.getPlacementById(id);
+        if(auth.employerId !== placement.employer_id) {
+            throw new SuperError(ERR_FORBIDDEN, 'You cannot close the applications for this placement');
+        }
         let result = await placementDAO.closePlacementById(id);
         if(result == 1){
             return "Your placement has been closed correctly."
